Добавление виджета жестов (iOS) (работает более отзывчивое управление 3D-моделью и добавленио вращение при касании за пределами кисти(как на Android)) (iteration 92)

+++ iosApp/MotoricaStart/Presentation/Widgets/WidgetsListViewModel.swift
@@ -29,7 +29,7 @@ protocol WidgetsListViewModelInput {
     func closeQueriesSuggestions()
     func didSelectItem(at index: Int)
     func requestInicializeInformation()
-    func setCustomGestureSettingsOpener(_ handler: @escaping () -> Void)
+    func setCustomGestureSettingsOpener(_ handler: @escaping (Int) -> Void)
 }
 
 protocol WidgetsListViewModelOutput {
@@ -53,7 +53,7 @@ final class DefaultWidgetsListViewModel: WidgetsListViewModel {
     private let searchWidgetsUseCase: SearchWidgetsUseCase
     private let actions: WidgetsListViewModelActions?
     private let bleManager: BleManagerKmm
-    private var customGestureSettingsOpener: (() -> Void)?
+    private var customGestureSettingsOpener: ((Int) -> Void)?
     
     var currentPage: Int = 0
     var totalPageCount: Int = 1
@@ -194,7 +194,7 @@ final class DefaultWidgetsListViewModel: WidgetsListViewModel {
         )
     }
     
-    func setCustomGestureSettingsOpener(_ handler: @escaping () -> Void) {
+    func setCustomGestureSettingsOpener(_ handler: @escaping (Int) -> Void) {
         customGestureSettingsOpener = handler
     }
 }
